# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduLearn EdTech Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  student: "student",
  trainer: "trainer",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  trainer: "trainer",
  student: "student",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  profilePhoto?: string | null;
  isBlocked: boolean;
  isApproved: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateUserBody {
  name?: string;
  profilePhoto?: string;
  isBlocked?: boolean;
  isApproved?: boolean;
}

export interface Course {
  id: number;
  title: string;
  description: string;
  category: string;
  trainerId: number;
  trainerName: string;
  thumbnail?: string | null;
  price: number;
  enrolledCount: number;
  rating: number;
  ratingCount: number;
  lessonCount: number;
  isPublished: boolean;
  createdAt: string;
}

export interface Lesson {
  id: number;
  courseId: number;
  title: string;
  videoUrl?: string | null;
  pdfUrl?: string | null;
  duration?: number | null;
  orderIndex: number;
  moduleTitle?: string | null;
}

export interface CourseRating {
  id: number;
  courseId: number;
  studentId: number;
  studentName: string;
  rating: number;
  review?: string | null;
  createdAt: string;
}

export type CourseDetail = Course & {
  lessons?: Lesson[];
  ratings?: CourseRating[];
};

export interface CreateCourseBody {
  title: string;
  description: string;
  category: string;
  thumbnail?: string;
  price: number;
  isPublished?: boolean;
}

export interface UpdateCourseBody {
  title?: string;
  description?: string;
  category?: string;
  thumbnail?: string;
  price?: number;
  isPublished?: boolean;
}

export interface CreateLessonBody {
  title: string;
  videoUrl?: string;
  pdfUrl?: string;
  duration?: number;
  moduleTitle?: string;
  orderIndex?: number;
}

export interface RateCourseBody {
  rating: number;
  review?: string;
}

export type EnrollmentCompletionStatus =
  (typeof EnrollmentCompletionStatus)[keyof typeof EnrollmentCompletionStatus];

export const EnrollmentCompletionStatus = {
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface Enrollment {
  id: number;
  studentId: number;
  courseId: number;
  courseTitle: string;
  courseThumbnail?: string | null;
  trainerName: string;
  progressPercentage: number;
  completionStatus: EnrollmentCompletionStatus;
  completedLessons: number[];
  enrollmentDate: string;
}

export interface EnrollBody {
  courseId: number;
}

export interface UpdateProgressBody {
  lessonId: number;
  completed: boolean;
}

export type AttendanceStatus =
  (typeof AttendanceStatus)[keyof typeof AttendanceStatus];

export const AttendanceStatus = {
  present: "present",
  absent: "absent",
} as const;

export interface Attendance {
  id: number;
  studentId: number;
  studentName: string;
  courseId: number;
  courseTitle: string;
  sessionDate: string;
  status: AttendanceStatus;
  createdAt: string;
}

export type MarkAttendanceBodyStatus =
  (typeof MarkAttendanceBodyStatus)[keyof typeof MarkAttendanceBodyStatus];

export const MarkAttendanceBodyStatus = {
  present: "present",
  absent: "absent",
} as const;

export interface MarkAttendanceBody {
  studentId: number;
  courseId: number;
  sessionDate: string;
  status: MarkAttendanceBodyStatus;
}

export type GeneratePlanBodySkillLevel =
  (typeof GeneratePlanBodySkillLevel)[keyof typeof GeneratePlanBodySkillLevel];

export const GeneratePlanBodySkillLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export type GeneratePlanBodyLearningStyle =
  (typeof GeneratePlanBodyLearningStyle)[keyof typeof GeneratePlanBodyLearningStyle];

export const GeneratePlanBodyLearningStyle = {
  videos: "videos",
  projects: "projects",
  theory: "theory",
  mixed: "mixed",
} as const;

export interface GeneratePlanBody {
  careerGoal: string;
  skillLevel: GeneratePlanBodySkillLevel;
  hoursPerDay: number;
  duration: number;
  preferredTechnologies?: string[];
  learningStyle?: GeneratePlanBodyLearningStyle;
}

export interface RoadmapWeek {
  week: number;
  weekLabel: string;
  topics: string[];
  recommendedCourse?: string | null;
  isCompleted: boolean;
}

export interface Milestone {
  id: number;
  title: string;
  description: string;
  targetWeek: number;
  isCompleted: boolean;
}

export interface LearningPlan {
  id: number;
  studentId: number;
  careerGoal: string;
  skillLevel: string;
  hoursPerDay: number;
  duration: number;
  roadmap: RoadmapWeek[];
  recommendedCourseIds: number[];
  milestones: Milestone[];
  createdAt: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface PlatformStats {
  totalUsers: number;
  totalTrainers: number;
  totalStudents: number;
  totalCourses: number;
  totalEnrollments: number;
  recentEnrollments: Enrollment[];
  topCourses: Course[];
  enrollmentsByCategory: CategoryCount[];
}

export interface TrainerStats {
  totalCourses: number;
  totalStudents: number;
  totalEnrollments: number;
  avgRating: number;
  courses: Course[];
  recentEnrollments: Enrollment[];
}

export interface StudentStats {
  enrolledCourses: number;
  completedCourses: number;
  totalHoursLearned: number;
  attendancePercentage: number;
  currentStreak: number;
  enrollments: Enrollment[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  courseCount: number;
}

export interface CreateCategoryBody {
  name: string;
  slug: string;
}

export type ListUsersParams = {
  role?: ListUsersRole;
  search?: string;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  admin: "admin",
  trainer: "trainer",
  student: "student",
} as const;

export type ListCoursesParams = {
  category?: string;
  search?: string;
  trainerId?: number;
};

export type ListAttendanceParams = {
  courseId?: number;
  studentId?: number;
};