# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EduLearn EdTech Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: Get attendance records
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark attendance (trainer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkAttendanceBody"
      responses:
        "201":
          description: Attendance marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /ai/generate-plan:
    post:
      operationId: generateLearningPlan
      tags: [ai]
      summary: Generate AI learning plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePlanBody"
      responses:
        "200":
          description: Generated learning plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningPlan"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, trainer]
      required:
        - name
        - email
        - password
        - role
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, trainer, student]
        profilePhoto:
          type: string
          nullable: true
        isBlocked:
          type: boolean
        isApproved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isBlocked
        - isApproved
        - createdAt
...
    Attendance:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        courseId:
          type: integer
        courseTitle:
          type: string
        sessionDate:
          type: string
        status:
          type: string
          enum: [present, absent]
        createdAt:
          type: string
      required:
        - id
        - studentId
        - studentName
        - courseId
        - courseTitle
        - sessionDate
        - status
        - createdAt
    MarkAttendanceBody:
      type: object
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
        sessionDate:
          type: string
        status:
          type: string
          enum: [present, absent]
      required:
        - studentId
        - courseId
        - sessionDate
        - status
    GeneratePlanBody:
      type: object
      properties:
        careerGoal:
          type: string
        skillLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        hoursPerDay:
          type: number
        duration:
          type: integer
        preferredTechnologies:
          type: array
          items:
            type: string
        learningStyle:
          type: string
          enum: [videos, projects, theory, mixed]
      required:
        - careerGoal
        - skillLevel
        - hoursPerDay
        - duration
    LearningPlan:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        careerGoal:
          type: string
        skillLevel:
          type: string
        hoursPerDay:
          type: number
        duration:
          type: integer
        roadmap:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapWeek"
        recommendedCourseIds:
          type: array
          items:
            type: integer
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        createdAt:
          type: string
      required:
        - id
        - studentId
        - careerGoal
        - skillLevel
        - hoursPerDay
        - duration
        - roadmap
        - recommendedCourseIds
        - milestones
        - createdAt
    RoadmapWeek:
      type: object
      properties:
        week:
          type: integer
        weekLabel:
          type: string
        topics:
          type: array
          items:
            type: string
        recommendedCourse:
          type: string
          nullable: true
        isCompleted:
          type: boolean
      required:
        - week
        - weekLabel
        - topics
        - isCompleted
    Milestone:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        targetWeek:
          type: integer
        isCompleted:
          type: boolean
      required:
        - id
        - title
        - description
        - targetWeek
        - isCompleted
...