# lib/api-zod/src/generated/types/markAttendanceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduLearn EdTech Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MarkAttendanceBodyStatus } from "./markAttendanceBodyStatus";

export interface MarkAttendanceBody {
  studentId: number;
  courseId: number;
  sessionDate: string;
  status: MarkAttendanceBodyStatus;
}