# lib/api-zod/src/generated/types/roadmapWeek.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EduLearn EdTech Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RoadmapWeek {
  week: number;
  weekLabel: string;
  topics: string[];
  recommendedCourse?: string | null;
  isCompleted: boolean;
}